test fix (plus a little refactoring)
