+++ tests/unit/test_pbreports_report_amplicon_analysis_input.py
@@ -6,6 +6,7 @@ import json
 from pprint import pformat
 
 from pbcommand.models.report import Report
+import pbcommand.testkit
 
 from pbreports.report.amplicon_analysis_input import run_to_report
 from base_test_case import LOCAL_DATA, run_backticks, validate_report_complete
@@ -53,3 +54,18 @@ class TestLongAmpliconAnalysisPcrReport(unittest.TestCase):
         # cleanup
         if os.path.exists(report_json):
             os.remove(report_json)
+
+
+class TestAmpliconAnalysisInput(pbcommand.testkit.PbTestApp):
+    name = 'amplicon_analysis_input.csv'
+    csv_file_name = os.path.join(LOCAL_DATA, "amplicon_analysis_input", name)
+    t = tempfile.NamedTemporaryFile(delete=False, suffix=".json")
+    t.close()
+    report_json = t.name
+    DRIVER_BASE = "python -m pbreports.report.amplicon_analysis_input "
+    DRIVER_EMIT = DRIVER_BASE + " --emit-tool-contract "
+    DRIVER_RESOLVE = DRIVER_BASE + " --resolved-tool-contract "
+    REQUIRES_PBCORE = False
+    INPUT_FILES = [csv_file_name]
+    OUTPUT_FILES = [report_json]
+    TASK_OPTIONS = {}
