+++ tests/unit/test_pbreports_report_amplicon_analysis_timing.py
@@ -6,6 +6,7 @@ import json
 from pprint import pformat
 
 from pbcommand.models.report import Report
+import pbcommand.testkit
 
 from pbreports.report.amplicon_analysis_timing import run_to_report
 from base_test_case import LOCAL_DATA, run_backticks, validate_report_complete
@@ -52,3 +53,20 @@ class TestIntegrationAmpliconAnalysisTimingReport(TestAmpliconAnalysisTimingRepo
         # cleanup
         if os.path.exists(report_json):
             os.remove(report_json)
+
+
+class TestAmpliconAnalysisTiming(pbcommand.testkit.PbTestApp):
+    DATA_DIR = os.path.join(LOCAL_DATA, "amplicon_analysis_timing")
+    log_file = os.path.join(DATA_DIR, "amplicon_analysis.log")
+    t = tempfile.NamedTemporaryFile(delete=False, suffix=".json")
+    t.close()
+    report_json = t.name
+    DRIVER_BASE = "python -m pbreports.report.amplicon_analysis_timing "
+    DRIVER_EMIT = DRIVER_BASE + " --emit-tool-contract "
+    DRIVER_RESOLVE = DRIVER_BASE + " --resolved-tool-contract "
+    REQUIRES_PBCORE = False
+    INPUT_FILES = [
+        log_file,
+    ]
+    OUTPUT_FILES = [report_json]
+    TASK_OPTIONS = {}
